+++ games_services_platform_interface/lib/models/access_point_location.dart
@@ -1,0 +1,3 @@
+enum AccessPointLocation {
+  topLeading, topTrailing, bottomLeading, bottomTrailing
+}